Iken Multisrc, move Vortex Scan and MangaGalaxy
